requirements.txt
streamlit
nltk
pandas
python-dotenv
scikit-learn
